Imported Upstream version 1.1.0

+++ modules/access/gnomevfs.c
@@ -2,7 +2,7 @@
  * gnomevfs.c: GnomeVFS input
  *****************************************************************************
  * Copyright (C) 2005 the VideoLAN team
- * $Id: b0ce8ecf020550be4be8e29e12f2a0c1e7d59030 $
+ * $Id: 93852102d2e8d438832e2376debb67bf312ec6ca $
  *
  * Authors: Benjamin Pracht <bigben -AT- videolan -DOT- org>
  *
@@ -24,6 +24,7 @@
 /*****************************************************************************
  * Preamble
  *****************************************************************************/
+
 #ifdef HAVE_CONFIG_H
 # include "config.h"
 #endif
@@ -36,7 +37,7 @@
 
 
 #include <vlc_charset.h>
-#include "vlc_url.h"
+#include <vlc_url.h>
 
 /*****************************************************************************
  * Module descriptor
@@ -64,9 +65,9 @@ vlc_module_end ()
 /*****************************************************************************
  * Exported prototypes
  *****************************************************************************/
-static int  Seek( access_t *, int64_t );
-static int  Read( access_t *, uint8_t *, size_t );
-static int  Control( access_t *, int, va_list );
+static int     Seek( access_t *, uint64_t );
+static ssize_t Read( access_t *, uint8_t *, size_t );
+static int     Control( access_t *, int, va_list );
 
 struct access_sys_t
 {
@@ -285,7 +286,7 @@ static void Close( vlc_object_t * p_this )
 /*****************************************************************************
  * Read: standard read on a file descriptor.
  *****************************************************************************/
-static int Read( access_t *p_access, uint8_t *p_buffer, size_t i_len )
+static ssize_t Read( access_t *p_access, uint8_t *p_buffer, size_t i_len )
 {
     access_sys_t *p_sys = p_access->p_sys;
     GnomeVFSFileSize i_read_len;
@@ -338,7 +339,7 @@ static int Read( access_t *p_access, uint8_t *p_buffer, size_t i_len )
 /*****************************************************************************
  * Seek: seek to a specific location in a file
  *****************************************************************************/
-static int Seek( access_t *p_access, int64_t i_pos )
+static int Seek( access_t *p_access, uint64_t i_pos )
 {
     access_sys_t *p_sys = p_access->p_sys;
     int i_ret;
